requirements.txt
streamlit
langchain
openai
faiss-cpu
pypdf
tiktoken
python-dotenv
langchain-community